feat: implement core mobile POS architecture with local database, domain logic, and presentation screens

+++ lib/data/local/daos/cloud_sync_event_dao.dart
@@ -88,7 +88,7 @@ class CloudSyncEventDao extends DatabaseAccessor<CloudDatabase>
   }
 
   Future<void> updateCursor(
-      String storeId, String deviceId, int cursor) async {
+      String storeId, String deviceId, BigInt cursor) async {
     final id = '${storeId}_$deviceId';
     await into(syncCursors).insertOnConflictUpdate(
       SyncCursorsCompanion.insert(
