+++ lib/data/local/daos/sync_event_dao.dart
@@ -97,7 +97,7 @@ class SyncEventDao extends DatabaseAccessor<AppDatabase>
 
   /// Upsert the sync cursor for a device.
   Future<void> updateCursor(
-      String storeId, String deviceId, int cursor) async {
+      String storeId, String deviceId, BigInt cursor) async {
     final id = '${storeId}_$deviceId';
     await into(syncCursors).insertOnConflictUpdate(
       SyncCursorsCompanion.insert(
